--- daycan-back/api/src/main/java/com/daycan/auth/AuthException.java
@@ -1,16 +1,0 @@
-package com.daycan.auth;
-
-import com.daycan.common.exception.ApplicationException;
-import com.daycan.common.response.status.AuthErrorStatus;
-
-public class AuthException extends ApplicationException {
-
-  public AuthException(AuthErrorStatus errorStatus) {
-    super(errorStatus);
-  }
-
-  public AuthException(AuthErrorStatus errorStatus, Object data) {
-    super(errorStatus, data);
-  }
-
-}
