+++ daycan-back/api/src/main/java/com/daycan/auth/dto/LoginResponse.java
@@ -1,4 +1,4 @@
-package com.daycan.auth.payload;
+package com.daycan.auth.dto;
 
 public record LoginResponse(
     String accessToken,
