+++ daycan-back/api/src/main/java/com/daycan/auth/dto/ReissueRequest.java
@@ -1,3 +1,3 @@
-package com.daycan.auth.payload;
+package com.daycan.auth.dto;
 
 public record ReissueRequest(String refreshToken) {}
