+++ daycan-back/api/src/main/java/com/daycan/auth/dto/Token.java
@@ -1,9 +1,6 @@
-package com.daycan.auth.token;
+package com.daycan.auth.dto;
 
-import com.daycan.auth.AuthException;
-import com.daycan.common.response.status.AuthErrorStatus;
 import java.util.Date;
-import lombok.Value;
 
 public record Token(String value, Date expiry) {
   public boolean isExpired() {
