+++ daycan-back/api/src/main/java/com/daycan/auth/exception/AuthException.java
@@ -1,0 +1,11 @@
+package com.daycan.auth.exception;
+
+import com.daycan.common.response.status.AuthErrorStatus;
+
+public class AuthException extends RuntimeException {
+
+  public AuthException(AuthErrorStatus errorStatus) {
+    super(errorStatus.getMessage());
+  }
+
+}
