+++ daycan-back/api/src/main/java/com/daycan/auth/model/AuthPrincipal.java
@@ -1,4 +1,4 @@
-package com.daycan.auth;
+package com.daycan.auth.model;
 
 import lombok.Getter;
