+++ daycan-back/api/src/main/java/com/daycan/auth/model/CenterPrincipal.java
@@ -1,4 +1,4 @@
-package com.daycan.auth;
+package com.daycan.auth.model;
 
 public class CenterPrincipal extends AuthPrincipal {
   public CenterPrincipal(String username) {
