+++ daycan-back/api/src/main/java/com/daycan/auth/model/TokenType.java
@@ -1,4 +1,4 @@
-package com.daycan.auth;
+package com.daycan.auth.model;
 
 public enum TokenType {
   ACCESS,
