+++ daycan-back/api/src/main/java/com/daycan/auth/repository/RefreshTokenRepository.java
@@ -1,6 +1,6 @@
 package com.daycan.auth.repository;
 
-import com.daycan.auth.token.RefreshToken;
+import com.daycan.auth.entity.RefreshToken;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
